backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
transformers==4.35.0
torch==2.1.0
openai==1.3.0
pydantic==2.4.0
soundfile==0.12.1
python-dotenv==1.0.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
python-multipart==0.0.6